pyproject.toml
[tool.poetry]
name = "data-analysis-agent"
version = "0.1.0"
description = ""
authors = ["Maitreya Satish Zalte <zaltesarita77@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain-openai = "^0.1.8"
langchain-experimental = "^0.0.59"
streamlit = "^1.35.0"
tabulate = "^0.9.0"
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
langchain-groq = "^0.1.5"
scikit-learn = "^1.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
